+++ cli-tools/wdio/lib/wdio-config/presets/default.js
@@ -44,5 +44,10 @@ exports.config = {
     global.expect = chai.expect;
     chai.config.includeStack = true;
     chai.should();
+  },
+  afterTest(test) {
+    // if (typeof test.error !== 'undefined') {
+    browser.takeScreenshot();
+    // }
   }
 };