+++ cli-tools/wdio/lib/wdio-config/wdio-presets.js
@@ -3,5 +3,7 @@ const path = require('path');
 module.exports.configMap = {
   'none': path.join(__dirname, 'presets', 'none'),
   'default': path.join(__dirname, 'presets', 'default.js'),
-  'sauce': path.join(__dirname, 'presets', 'sauce.js')
+  'sauce': path.join(__dirname, 'presets', 'sauce.js'),
+  'testingbot': path.join(__dirname, 'presets', 'testingbot.js'),
+  'crossbrowser': path.join(__dirname, 'presets', 'crossbrowser.js')
 };